Integrate with receptoRA
Whoops. Forgot to include receptoRA last commit. It's here, plus:
Full integration! A second app, receptoRA needs to be launched and take the saved data to analyze it.
What will still be needed is a data storage structure to permit people to come back and look at their results later.